pyproject.toml
[project]
name = "telegram"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "greenlet>=3.5.0",
    "httpx[http2]>=0.28.1",
    "psycopg[binary]>=3.3.4",
    "pydantic>=2.13.3",
    "python-telegram-bot>=22.7",
    "smolagents[audio,mlx-lm,openai,toolkit,vision]>=1.24.0",
    "sqlmodel>=0.0.38",
    "telegramify-markdown>=1.1.2",
]
